requirements.txt
discord.py
pillow
markdown-pdf
requests
aiohttp[speedups]